+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Summary: The alphabet module is almost *stable*. Several things in meta and rang
 * Added `bio::ranges::back_insertable` and `bio::ranges::back_insertable_with` as light-weight "container" concepts.
 * Added `bio::views::char_strictly_to` and `bio::views::validate_char_for`; as well as `bio::views::char_conversion_view_t`.
 * Added `bio::views::transform_by_pos`, a more flexible version of `std::views::transform`.
+* Added `bio::views::add_reverse_complement` which turns a range of sequences into range of sequences with their reverse complements interleaved as every second element.
 * All alphabet literals (e.g. 'C'_dna4, "GATTACA"_dna4) are now strict and validated at compile-time. Character literals are `consteval`. This prevents bugs early on.
 
 ## Bug-fixes
@@ -64,4 +65,3 @@ Summary: The alphabet module is almost *stable*. Several things in meta and rang
 
 * This release is not yet API-stable.
 * This release is not backwards compatible with any BioC++ or SeqAn version.
-